+++ apps/tangle-dapp/types/index.ts
@@ -197,3 +197,8 @@ export type ServiceParticipant = {
 export enum NetworkFeature {
   Faucet,
 }
+
+export enum ExplorerType {
+  Substrate = 'polkadot',
+  EVM = 'web3',
+}
